BF-37 remove warnings in Password Reset

+++ backend/src/main/java/com/brainfood/security/PasswordResetManager.java
@@ -51,7 +51,9 @@ public class PasswordResetManager {
     public Response changePassword(String tokenStr, String password) {
         PasswordResetToken token = passwordResetTokenRepository.findByToken(tokenStr);
         if(token == null) return Response.InvalidToken;
-        UserCredentials user = passwordResetTokenRepository.findById(tokenStr).get().getUser();
+        token =passwordResetTokenRepository.findById(tokenStr).orElse(null);
+        UserCredentials user =token!=null? token.getUser():null;
+        if(user == null) return Response.InvalidToken;
         user.setPassword(userAuthenticator.bCryptPasswordEncoder.encode(password));
         passwordResetTokenRepository.deleteById(tokenStr);
         return Response.PasswordChanged;
